handle some features in Setting fragment like search by City , so i handle Map Fragment to search by city , and finish the feature that add the city to favourite and i can delete it frm favourite

+++ app/src/main/java/com/aser/weatherwhisper/model/Minutely.kt
@@ -1,6 +1,6 @@
 package com.aser.weatherwhisper.model
 
 data class Minutely(
-    val dt: Int,
-    val precipitation: Int
+    val dt: Long,
+    val precipitation: Double
 )
